requirements.txt
Django==2.1.5
psycopg2
djangorestframework
requests
nltk
stop_words